requirement.txt
PyYAML==6.0
requests==2.28.2
bs4==0.0.1
pandas==1.5.3
lxml==4.9.2
mysql-connector-python==8.0.32